requirements.txt
discord~=2.3.2
requests==2.32.3
python-dotenv==1.0.1